backend/requirements.txt
# ─── CORE API & ORCHESTRATION ───
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
httpx==0.27.0

# ─── DATABASE & QUEUE ───
asyncpg==0.29.0
redis==5.0.3

# ─── WEBRTC & VIDEO INGESTION ───
livekit-api==0.5.0
livekit-agents==0.5.0

# ─── AI & VISION SWARM ───
opencv-python-headless==4.9.0.80
numpy==1.26.4
tritonclient[grpc]==2.43.0
ultralytics==8.1.24  # Used ONLY for exporting the model, not inference

# ─── SECURITY ───
PyJWT==2.8.0